Desktop/projects/iampersonalAI/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==12.0
firebase-admin==6.5.0
python-jose[cryptography]==3.3.0
pydantic==2.9.0
pydantic-settings==2.5.0
aiohttp==3.10.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
python-multipart==0.0.9
structlog==24.4.0
